requirements.txt
awscli-local==0.20
awscli==1.27.44
boto3==1.26.43
botocore==1.29.44
jmespath==1.0.1
launchpadlib==1.10.13
localstack-client==1.39
psycopg2-binary==2.9.5
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
requests==2.31.0
subprocess.run==0.0.8